Fix tests; Remove /get_forks/, /get_registrations/, and /get_readable_descendants
And add _view_project tests for embedding forks/registrations/descendants

Also, annotate queryset for forks and registration with nlogs
