+++ tests/test_views.py
@@ -775,36 +775,6 @@ class TestProjectViews(OsfTestCase):
         res = self.app.delete_json(url, {'tag': 'troz'}, auth=self.auth, expect_errors=True)
         assert_equal(res.status_code, http.CONFLICT)
 
-    # Regression test for https://github.com/CenterForOpenScience/osf.io/issues/1478
-    @mock.patch('website.archiver.tasks.archive')
-    def test_registered_projects_contributions(self, mock_archive):
-        # register a project
-        self.project.register_node(get_default_metaschema(), Auth(user=self.project.creator), '', None)
-        # get the first registered project of a project
-        url = self.project.api_url_for('get_registrations')
-        res = self.app.get(url, auth=self.auth)
-        data = res.json
-        pid = data['nodes'][0]['id']
-        url2 = api_url_for('get_summary', pid=pid)
-        # count contributions
-        res2 = self.app.get(url2, auth=self.auth)
-        data = res2.json
-        assert_is_not_none(data['summary']['nlogs'])
-
-    def test_forks_contributions(self):
-        # fork a project
-        self.project.fork_node(Auth(user=self.project.creator))
-        # get the first forked project of a project
-        url = self.project.api_url_for('get_forks')
-        res = self.app.get(url, auth=self.auth)
-        data = res.json
-        pid = data['nodes'][0]['id']
-        url2 = api_url_for('get_summary', pid=pid)
-        # count contributions
-        res2 = self.app.get(url2, auth=self.auth)
-        data = res2.json
-        assert_is_not_none(data['summary']['nlogs'])
-
     def test_remove_project(self):
         url = self.project.api_url
         res = self.app.delete_json(url, {}, auth=self.auth).maybe_follow()
@@ -931,27 +901,6 @@ class TestProjectViews(OsfTestCase):
         self.project.reload()
         assert_equal(self.project.title, 'newtitle')
 
-    # Regression test
-    def test_get_registrations_sorted_by_registered_date_descending(self):
-        # register a project several times, with various registered_dates
-        registrations = []
-        for days_ago in (21, 3, 2, 8, 13, 5, 1):
-            registration = RegistrationFactory(project=self.project)
-            reg_date = registration.registered_date - dt.timedelta(days_ago)
-            registration.registered_date = reg_date
-            registration.save()
-            registrations.append(registration)
-
-        registrations.sort(key=lambda r: r.registered_date, reverse=True)
-        expected = [r._id for r in registrations]
-
-        registrations_url = self.project.api_url_for('get_registrations')
-        res = self.app.get(registrations_url, auth=self.auth)
-        data = res.json
-        actual = [n['id'] for n in data['nodes']]
-
-        assert_equal(actual, expected)
-
 
 class TestEditableChildrenViews(OsfTestCase):
 
@@ -1001,150 +950,6 @@ class TestEditableChildrenViews(OsfTestCase):
         assert_equal(self.project_results['children'][1]['title'], self.grandchild.title)
         assert_equal(self.project_results['children'][2]['title'], self.great_grandchild.title)
         assert_equal(self.project_results['children'][3]['title'], self.great_great_grandchild.title)
-
-
-class TestChildrenViews(OsfTestCase):
-
-    def setUp(self):
-        OsfTestCase.setUp(self)
-        self.user = AuthUserFactory()
-
-    def test_get_readable_descendants(self):
-        project = ProjectFactory(creator=self.user)
-        child = NodeFactory(parent=project, creator=self.user)
-
-        url = project.api_url_for('get_readable_descendants')
-        res = self.app.get(url, auth=self.user.auth)
-
-        nodes = res.json['nodes']
-        assert_equal(len(nodes), 1)
-        assert_equal(nodes[0]['id'], child._primary_key)
-
-    def test_get_readable_descendants_includes_pointers(self):
-        project = ProjectFactory(creator=self.user)
-        pointed = ProjectFactory()
-        node_relation = project.add_pointer(pointed, auth=Auth(self.user))
-        project.save()
-
-        url = project.api_url_for('get_readable_descendants')
-        res = self.app.get(url, auth=self.user.auth)
-
-        nodes = res.json['nodes']
-        assert_equal(len(nodes), 1)
-        assert_equal(nodes[0]['title'], pointed.title)
-        assert_equal(nodes[0]['id'], node_relation._id)
-
-    def test_readable_descendants_masked_by_permissions(self):
-        # Users should be able to see through components they do not have
-        # permissions to.
-        # Users should not be able to see through links to nodes they do not
-        # have permissions to.
-        #
-        #                   1(AB)
-        #                  /  |  \
-        #                 *   |   \
-        #                /    |    \
-        #             2(A)  4(B)    7(A)
-        #               |     |     |    \
-        #               |     |     |     \
-        #             3(AB) 5(B)    8(AB) 9(B)
-        #                     |
-        #                     |
-        #                   6(A)
-        #
-        #
-        userA = AuthUserFactory(fullname='User A')
-        userB = AuthUserFactory(fullname='User B')
-
-        project1 = ProjectFactory(creator=self.user, title='One')
-        project1.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-        project1.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        component2 = ProjectFactory(creator=self.user, title='Two')
-        component2.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-
-        component3 = ProjectFactory(creator=self.user, title='Three')
-        component3.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-        component3.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        component4 = ProjectFactory(creator=self.user, title='Four')
-        component4.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        component5 = ProjectFactory(creator=self.user, title='Five')
-        component5.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        component6 = ProjectFactory(creator=self.user, title='Six')
-        component6.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-
-        component7 = ProjectFactory(creator=self.user, title='Seven')
-        component7.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-
-        component8 = ProjectFactory(creator=self.user, title='Eight')
-        component8.add_contributor(userA, auth=Auth(self.user), permissions=['read'])
-        component8.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        component9 = ProjectFactory(creator=self.user, title='Nine')
-        component9.add_contributor(userB, auth=Auth(self.user), permissions=['read'])
-
-        project1.add_pointer(component2, Auth(self.user))
-        NodeRelation.objects.create(parent=project1, child=component4)
-        NodeRelation.objects.create(parent=project1, child=component7)
-        NodeRelation.objects.create(parent=component2, child=component3)
-        NodeRelation.objects.create(parent=component4, child=component5)
-        NodeRelation.objects.create(parent=component5, child=component6)
-        NodeRelation.objects.create(parent=component7, child=component8)
-        NodeRelation.objects.create(parent=component7, child=component9)
-
-        url = project1.api_url_for('get_readable_descendants')
-
-        res = self.app.get(url, auth=userA.auth)
-        assert_equal(len(res.json['nodes']), 3)
-        for node in res.json['nodes']:
-            assert_in(node['title'], ['Two', 'Six', 'Seven'])
-
-        res = self.app.get(url, auth=userB.auth)
-        assert_equal(len(res.json['nodes']), 3)
-        for node in res.json['nodes']:
-            assert_in(node['title'], ['Four', 'Eight', 'Nine'])
-
-    def test_get_readable_descendants_filter_for_permissions(self):
-        # self.user has admin access to this project
-        project = ProjectFactory(creator=self.user)
-
-        # self.user only has read access to this project, which project points
-        # to
-        read_only_pointed = ProjectFactory()
-        read_only_creator = read_only_pointed.creator
-        read_only_pointed.add_contributor(self.user, auth=Auth(read_only_creator), permissions=['read'])
-        read_only_pointed.save()
-
-        # self.user only has read access to this project, which is a subproject
-        # of project
-        read_only = ProjectFactory()
-        read_only_pointed.add_contributor(self.user, auth=Auth(read_only_creator), permissions=['read'])
-        NodeRelation.objects.create(parent=project, child=read_only)
-
-        # self.user adds a pointer to read_only
-        project.add_pointer(read_only_pointed, Auth(self.user))
-        project.save()
-
-        url = project.api_url_for('get_readable_descendants')
-        res = self.app.get(url, auth=self.user.auth)
-        assert_equal(len(res.json['nodes']), 2)
-
-        url = project.api_url_for('get_readable_descendants', permissions='write')
-        res = self.app.get(url, auth=self.user.auth)
-        assert_equal(len(res.json['nodes']), 0)
-
-    def test_get_readable_descendants_render_nodes_receives_auth(self):
-        project = ProjectFactory(creator=self.user)
-        NodeFactory(parent=project, creator=self.user)
-
-        url = project.api_url_for('get_readable_descendants')
-        res = self.app.get(url, auth=self.user.auth)
-
-        perm = res.json['nodes'][0]['permissions']
-        assert_equal(perm, 'admin')
 
 
 class TestGetNodeTree(OsfTestCase):
@@ -4395,16 +4200,6 @@ class TestForkViews(OsfTestCase):
         url = self.project.api_url_for('node_fork_page')
         res = self.app.post_json(url, auth=contributor.auth)
         assert_equal(res.status_code, 200)
-
-    def test_registered_forks_dont_show_in_fork_list(self):
-        fork = self.project.fork_node(self.consolidated_auth)
-        RegistrationFactory(project=fork)
-
-        url = self.project.api_url_for('get_forks')
-        res = self.app.get(url, auth=self.user.auth)
-
-        assert_equal(len(res.json['nodes']), 1)
-        assert_equal(res.json['nodes'][0]['id'], fork._id)
 
 
 class TestProjectCreation(OsfTestCase):
